PostCSS: parser fix to support '&' nesting sign in any place of selector list, related nesting inspection and delete '&' quick fix changes

+++ postcss/src/org/intellij/plugins/postcss/actions/PostCssDeleteAmpersandQuickFix.java
@@ -4,7 +4,11 @@ import com.intellij.codeInspection.LocalQuickFixBase;
 import com.intellij.codeInspection.ProblemDescriptor;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.css.CssElementFactory;
+import com.intellij.psi.css.CssSelectorList;
+import com.intellij.psi.util.PsiTreeUtil;
 import org.intellij.plugins.postcss.PostCssBundle;
+import org.intellij.plugins.postcss.PostCssLanguage;
 import org.intellij.plugins.postcss.psi.PostCssPsiUtil;
 import org.jetbrains.annotations.NotNull;
 
@@ -16,8 +20,11 @@ public class PostCssDeleteAmpersandQuickFix extends LocalQuickFixBase {
   @Override
   public void applyFix(@NotNull Project project, @NotNull ProblemDescriptor descriptor) {
     PsiElement element = descriptor.getStartElement();
-    if (PostCssPsiUtil.isAmpersand(element)){
+    CssSelectorList selectorList = PsiTreeUtil.getParentOfType(element, CssSelectorList.class);
+    if (PostCssPsiUtil.isAmpersand(element) && selectorList != null) {
       element.delete();
+      selectorList.replace(
+        CssElementFactory.getInstance(element.getProject()).createSelectorList(selectorList.getText(), PostCssLanguage.INSTANCE));
     }
   }
 }
